Decoupled exercise:
this exercise focuses on writing an aspect similar to the Java "shared" aspect, called "decoupled". 
A decoupled object is an object that guarantees that every call to its public methods is executed in a dedicated thread, granting at least 1 thread per call. 
However, return values of the calls are still guaranteed to be returned in synchronous.